Add apiTokens, isValidAPIToken, reorgs tests a little
