+++ Tests/AppTests/AppTests.swift
@@ -20,15 +20,6 @@ import XCTest
 
 class AppTests: AppTestCase {
 
-    func test_Filemanager_checkoutsDirectory() throws {
-        Current.fileManager = .live
-        unsetenv("CHECKOUTS_DIR")
-        XCTAssertEqual(Current.fileManager.checkoutsDirectory(),
-                       DirectoryConfiguration.detect().workingDirectory + "SPI-checkouts")
-        setenv("CHECKOUTS_DIR", "/tmp/foo", 1)
-        XCTAssertEqual(Current.fileManager.checkoutsDirectory(), "/tmp/foo")
-    }
-
     func test_migrations() async throws {
         try await XCTAssertNoThrowAsync(try await app.autoRevert())
         try await XCTAssertNoThrowAsync(try await app.autoMigrate())
